Powerup Bugs Fixed
-Bug where second powerup touched causes a crash is fixed
-Bug where powerup effects don't apply fixed

+++ Blackout/Blackout/Level.cs
@@ -123,8 +123,8 @@ namespace Blackout
                     
                 }
             }
-            double tempYChange = changeY;
-            double tempXChange = changeX;
+            int tempYChange = (int)changeY;
+            int tempXChange = (int)changeX;
             if (mortimerMovesInX || hitATileWallX) {
                 tempXChange = 0;
             }
